[MCUX-83760] Fix MISRA C-2012 Rule 10.3
Implicit conversion of count from essential type signed 32-bit int to different or narrower essential type unsigned 32-bit int.
Explicitly convert count to ULONG

+++ lib/rpmsg_lite/porting/environment/rpmsg_env_threadx.c
@@ -377,7 +377,7 @@ int32_t env_create_mutex(void **lock, int32_t count)
         return -1;
     }
 
-    if (TX_SUCCESS == tx_semaphore_create((TX_SEMAPHORE *)semaphore_ptr, NULL, count))
+    if (TX_SUCCESS == tx_semaphore_create((TX_SEMAPHORE *)semaphore_ptr, NULL, (ULONG)count))
     {
         *lock = (void *)semaphore_ptr;
         return 0;
